+++ src/Services/calculator/functions/IC.function.ts
@@ -1,7 +1,8 @@
-import { QuarterlyData, PeriodicData } from "@/resources/consumers/price-check-consumer/discount-manager/discount-manager.typings";
+import { QuarterlyData } from "@/resources/consumers/price-check-consumer/discount-manager/discount-manager.typings";
 import { processPeriodicDatasets, annualizeByLastQuarter } from "@/resources/consumers/price-check-consumer/discount-manager/discount-manager.utils";
 import { TimePeriod } from "../calculator.typings";
 import AbstractFunction from "./AbstractFunction";
+import { PeriodicData } from "@/src/types";
 
 
 class IcFunction extends AbstractFunction {
