+++ src/Services/calculator/functions/PE.function.ts
@@ -5,8 +5,9 @@ import { PriceData, HistoricalPriceInput } from "@/services/historical-price/his
 import { historicalPriceService } from "../../../bootstrap";
 import { TimePeriod } from "../calculator.typings";
 import { days_between } from "../calculator.utils";
-import { QuarterlyData, PeriodicData } from "@/resources/consumers/price-check-consumer/discount-manager/discount-manager.typings";
+import { QuarterlyData } from "@/resources/consumers/price-check-consumer/discount-manager/discount-manager.typings";
 import { annualizeByAdd } from "@/resources/consumers/price-check-consumer/discount-manager/discount-manager.utils";
+import { PeriodicData } from "@/src/types";
