+++ src/Services/calculator/functions/ROIC.function.ts
@@ -1,8 +1,9 @@
-import { QuarterlyData, PeriodicData } from "@/resources/consumers/price-check-consumer/discount-manager/discount-manager.typings";
+import { QuarterlyData } from "@/resources/consumers/price-check-consumer/discount-manager/discount-manager.typings";
 import { processPeriodicDatasets, annualizeByAdd } from "@/resources/consumers/price-check-consumer/discount-manager/discount-manager.utils";
 import { calculatorService } from "../../../bootstrap";
 import { TimePeriod } from "../calculator.typings";
 import AbstractFunction from "./AbstractFunction";
+import { PeriodicData } from "@/src/types";
 
 
 class RoicFunction extends AbstractFunction {
