+++ src/index.ts
@@ -3,8 +3,8 @@ import 'module-alias/register';
 import App from './app';
 import initializeEnv from '@/utils/initializeEnv';
 import ListenerController from './resources/controllers/listener.controller';
+import PriceCheckConsumer from '@/resources/consumers/price-check-consumer/price-check.consumer';
 import bootstrap from './bootstrap';
-import PriceCheckConsumer from './resources/consumers/price-check-consumer/price-check.consumer';
 
 
 initializeEnv().then(() => {
