--- src/resources/consumers/consumer.typings.ts
@@ -1,3 +1,0 @@
-export default interface Consumer {
-    startPolling(): Promise<void>   
-}
