+++ src/resources/consumers/price-check-consumer/discount-manager/discount-manager.typings.ts
@@ -1,11 +1,4 @@
-export type Period = 'Q1' | 'Q2' | 'Q3' | 'Q4' | 'FY';
-
-export interface PeriodicData {
-    cik: string
-    announcedDate: Date
-    period: Period
-    value: number
-}
+import { PeriodicData } from "@/src/types"
 
 export interface QuarterlyData {
     quarterlyShareholderEquity: PeriodicData[]
