--- src/resources/controllers/controller.typings.ts
@@ -1,8 +1,0 @@
-import { Router } from "express";
-
-interface Controller {
-    path: string, 
-    router: Router
-}
-
-export default Controller
