+++ src/resources/controllers/listener.controller.ts
@@ -3,8 +3,8 @@ import CONSTANTS from "../resource.contants";
 import HttpException from "@/utils/exceptions/HttpException";
 import { discountService, historicalPriceService } from "../../bootstrap";
 import { SimpleDiscount } from "@/services/discount/discount.typings";
-import Controller from "./controller.typings";
 import { checkSimpleDiscountIsOnSale } from "../resource.utils";
+import { Controller } from "@/src/types";
 
 class ListenerController implements Controller {
