+++ tsconfig.json
@@ -34,7 +34,7 @@
       "@/utils/*" : ["utils/*"],
       "@/middleware/*" : ["middleware/*"],
       "@/services/*" : ["services/*"],
-      "@/boostrap/*" : ["boostrap"]
+      "@/src/*" : ["./*"]
     },                                         /* Specify a set of entries that re-map imports to additional lookup locations. */
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     // "typeRoots": [],                                  /* Specify multiple folders that act like './node_modules/@types'. */
